@@ -4,8 +4,6 @@
   "author": "ExtensionEngine <info@extensionengine.com>",
   "license": "MIT",
   "version": "0.0.1",
-  "main": "dist/tce-scorm.js",
-  "module": "dist/tce-scorm.esm.js",
   "keywords": [],
   "homepage": "https://github.com/ExtensionEngine/tce-scorm#readme",
   "bugs": {
@@ -15,48 +13,70 @@
     "type": "git",
     "url": "git+https://mirodojkic@github.com/ExtensionEngine/tce-scorm.git"
   },
+  "files": [
+    "dist"
+  ],
+  "main": "./dist/tce-scorm.js",
+  "umd:main": "./dist/tce-scorm.umd.js",
+  "module": "./dist/tce-scorm.mjs",
+  "exports": {
+    ".": {
+      "import": "./dist/tce-scorm.mjs",
+      "require": "./dist/tce-scorm.js"
+    },
+    "./tce-scorm.css": "./dist/tce-scorm.css"
+  },
   "scripts": {
     "prebuild": "del-cli dist",
-    "build": "bili",
-    "dev": "bili --watch",
+    "build": "vite build",
+    "serve": "vite serve",
     "lint": "npm run lint:js && npm run lint:scss",
     "lint:js": "eslint --ext .js,.vue .",
     "lint:scss": "stylelint \"**/*.vue\" \"**/*.scss\""
   },
   "dependencies": {
-    "fast-xml-parser": "^3.17.5",
+    "fast-xml-parser": "^4.2.4",
     "lodash": "^4.17.21",
-    "p-min-delay": "^4.0.1"
+    "p-min-delay": "^4.0.2",
+    "vue": "^3.3.4"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.18.2",
-    "@babel/eslint-plugin": "^7.17.7",
-    "@extensionengine/eslint-config": "^3.0.1",
+    "@babel/eslint-plugin": "^7.22.5",
     "@extensionengine/rollup-plugin-tailor-ce": "^0.4.0",
-    "@extensionengine/stylelint-config": "^1.1.1",
-    "@rollup/plugin-alias": "^3.1.9",
-    "@vue/compiler-sfc": "^3.2.37",
-    "babel-eslint": "^10.1.0",
-    "bili": "^5.0.5",
-    "del-cli": "^4.0.1",
-    "eslint": "^8.19.0",
-    "eslint-config-semistandard": "^16.0.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.26.0",
+    "@rollup/plugin-alias": "^5.0.0",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vue/compiler-sfc": "^3.3.4",
+    "del-cli": "^5.0.0",
+    "eslint": "^8.43.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-config-semistandard": "^17.0.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-import-resolver-alias": "^1.1.2",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-n": "^15.2.4",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-vue": "9.2.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-require-sort": "^1.3.0",
+    "eslint-plugin-vue": "9.15.0",
+    "eslint-scope": "^7.2.0",
+    "postcss-html": "^1.5.0",
+    "postcss-scss": "^4.0.6",
+    "rollup": "^3.25.1",
     "rollup-plugin-copy": "^3.4.0",
-    "rollup-plugin-visualizer": "^4.2.0",
-    "rollup-plugin-vue": "5.1.9",
-    "sass": "^1.53.0",
-    "stylelint": "^13.8.0",
-    "vue-runtime-helpers": "^1.1.2",
-    "vue-template-compiler": "^2.7.4"
+    "rollup-plugin-postcss": "^4.0.2",
+    "rollup-plugin-visualizer": "^5.9.2",
+    "rollup-plugin-vue": "6.0.0",
+    "sass": "^1.63.4",
+    "stylelint": "^15.8.0",
+    "stylelint-config-css-modules": "^4.2.0",
+    "stylelint-config-recess-order": "^4.2.0",
+    "stylelint-config-recommended-vue": "^1.4.0",
+    "stylelint-config-standard-scss": "^9.0.0",
+    "vue-eslint-parser": "^9.3.1",
+    "vue-runtime-helpers": "^1.1.2"
   },
   "peerDependencies": {
-    "vee-validate": "^3.4.5",
-    "vuetify": "^2.3.21"
+    "vee-validate": "^4.9.6",
+    "vuetify": "^3.3.2"
   },
   "tailor": {
     "label": "Scorm",
