@@ -8,7 +8,8 @@
       name="SCORM component"
       icon="mdi-package-up"
       active-placeholder="Use toolbar to upload the SCORM package"
-      active-icon="mdi-arrow-up" />
+      active-icon="mdi-arrow-up"
+    />
     <div v-else>
       <preview-overlay v-if="!isDisabled && !isFocused" />
       <div class="container">
@@ -16,34 +17,34 @@
           :src="launchUrl"
           :tabindex="isDisabled || isFocused ? 0 : -1"
           frameborder="0"
-          class="element">
-        </iframe>
+          class="element"
+        />
       </div>
     </div>
   </div>
 </template>
 
-<script>
+<script setup>
+import { computed, inject, onMounted } from 'vue';
 import ElementPlaceholder from '@/tce-core/ElementPlaceholder.vue';
 import PreviewOverlay from '@/tce-core/PreviewOverlay.vue';
 
-export default {
-  name: 'tce-scorm',
-  inject: ['$elementBus'],
-  props: {
-    element: { type: Object, required: true },
-    isFocused: { type: Boolean, default: false },
-    isDisabled: { type: Boolean, default: false },
-    dense: { type: Boolean, default: false }
-  },
-  computed: {
-    launchUrl: vm => vm.element.data.launchUrl
-  },
-  mounted() {
-    this.$elementBus.on('save', data => this.$emit('save', data));
-  },
-  components: { ElementPlaceholder, PreviewOverlay }
-};
+const props = defineProps({
+  element: { type: Object, required: true },
+  isFocused: { type: Boolean, default: false },
+  isDisabled: { type: Boolean, default: false },
+  dense: { type: Boolean, default: false }
+});
+
+const emit = defineEmits(['save']);
+
+const $elementBus = inject('$elementBus');
+
+const launchUrl = computed(() => props.element.data.launchUrl);
+
+onMounted(() => {
+  $elementBus.on('save', data => emit('save', data));
+});
 </script>
 
 <style lang="scss" scoped>
