@@ -1,25 +1,33 @@
 import loader from './loader';
+import { ref } from 'vue';
 
-export default {
-  inject: ['$storageService'],
-  data: () => ({ uploading: false }),
-  methods: {
-    createFileForm(e) {
-      this.form = new FormData();
-      const [file] = e.target.files;
-      if (!file) return;
-      this.form.append('file', file, file.name);
-      this.form.append('unpack', true);
-    },
-    upload: loader(function (repositoryId, e) {
-      this.createFileForm(e);
-      return this.$storageService.upload(repositoryId, this.form)
-        .then(data => {
-          const { name } = this.form.get('file');
-          this.$emit('upload', { ...data, name });
-        }).catch(() => {
-          this.error = 'An error has occurred!';
-        });
-    }, 'uploading')
+export default function useUpload($storageService) {
+  const uploading = ref(false);
+
+  function createFileForm(e) {
+    const form = new FormData();
+    const [file] = e.target.files;
+    if (!file) return;
+    form.append('file', file, file.name);
+    form.append('unpack', true);
+    return form;
   }
-};
+
+  const upload = loader(function (repositoryId, e) {
+    const form = createFileForm(e);
+    return $storageService
+      .upload(repositoryId, form)
+      .then(data => {
+        const { name } = form.get('file');
+        return { ...data, name };
+      })
+      .catch(() => {
+        throw new Error('An error has occurred!');
+      });
+  }, uploading);
+
+  return {
+    uploading,
+    upload
+  };
+}
